refactor: 테스트 패키지를 도메인별로 분리하라
