+++ src/test/java/kitchenpos/FakeKitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos;
 
 import kitchenpos.infra.KitchenridersClient;
